Rename url data function to avoid referencing remote
This doesn’t necessarily need to be a remote URL. We could be fetching from an internal URL using the same mechanic so let’s not confuse people/things.
