+++ packages/core-data/src/fetch/index.js
@@ -1,2 +1,2 @@
 export { default as __experimentalFetchLinkSuggestions } from './__experimental-fetch-link-suggestions';
-export { default as __experimentalFetchRemoteUrlData } from './__experimental-fetch-remote-url-data';
+export { default as __experimentalFetchUrlData } from './__experimental-fetch-url-data';
